Move repo-tool.js to utils directory
